drpcclient: create a clientconn wrapping drpc.Conn
Here, we are updating constructor of clientConn to take in a
`drpc.Conn` type object instead of a supplier of `drpc.Conn`.
This simplifies the usage/construction of clientConn, and moves
the responsibility of invoking a supplier/dialer to the caller of
constructor. This way, closing of drpcpool would be a clear duty
of the caller of this constructor.