makefile: add gen-bazel target (#15)
This change adds a `gen-bazel` target which generates a `WORKSPACE` file
and runs gazelle to generate `BUILD.bazel` files. Generating these files
allows a local clone of the repository to be used directly when building
cockroach using the `override_repository` flag, for example:

```
./dev build short -- --override_repository=io_storj_drpc=${CURDIR}
```

The new files are added to `.gitignore` so that generating them doesn't
interfere with commits.

This also adds a `clean-bazel` target that cleans up these files.

Inspired by:
https://github.com/cockroachdb/pebble/commit/cab9f08de0db22259b8c58ebc0980aec7b7eaae0